Fixed rollbacks after an error during migration
ref https://github.com/TryGhost/knex-migrator/issues/241

By catching migration errors at the individual level, we're able to only
rollback the versions which have already been migrated. Updating the
migrate method to use async/await meant that it was easier to follow the
flow control and refactor the error handling to allow us to correctly
rollback.

The eslint file has had it's parser options updated manually because the
latest ghost-eslint-plugin causes a large number of errors, which fixing
would be outside the scope of this issue.

The logging is also improved to make it clear what is happening, when
migrations are starting, and also when rollbacks are being run.